# openapi/payops-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayOps Merchant API",
    "version": "0.3.0",
    "description": "Organization-scoped Solana stablecoin invoicing with hosted exact-amount Solana Pay checkout and finalized payment status."
  },
  "servers": [{ "url": "https://api.example.com" }],
  "security": [{ "sessionCookie": [] }, { "apiKey": [] }],
  "paths": {
...
    "/v1/exceptions": {
      "get": {
        "operationId": "listPaymentExceptions",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" },
          {
            "name": "state",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ExceptionReviewState" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment exception page",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["data", "nextCursor"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentException"
                      }
                    },
                    "nextCursor": { "type": ["string", "null"] }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/exceptions/{exceptionId}/history": {
      "get": {
        "operationId": "getPaymentExceptionHistory",
        "parameters": [{ "$ref": "#/components/parameters/ExceptionId" }],
        "responses": {
          "200": {
            "description": "Append-only exception history",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ExceptionCaseEvent"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/exceptions/{exceptionId}/assign": {
      "post": {
        "operationId": "assignPaymentException",
        "parameters": [
          { "$ref": "#/components/parameters/ExceptionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssignExceptionInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exception assigned",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentException" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "409": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/exceptions/{exceptionId}/resolve": {
      "post": {
        "operationId": "resolvePaymentException",
        "parameters": [
          { "$ref": "#/components/parameters/ExceptionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResolveExceptionInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exception resolved",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentException" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "409": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/exceptions/{exceptionId}/investigate": {
      "post": {
        "operationId": "startPaymentExceptionInvestigation",
        "parameters": [
          { "$ref": "#/components/parameters/ExceptionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionExceptionInput"
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PaymentException" },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "409": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/exceptions/{exceptionId}/escalate": {
      "post": {
        "operationId": "escalatePaymentException",
        "parameters": [
          { "$ref": "#/components/parameters/ExceptionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionExceptionInput"
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PaymentException" },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "409": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/exceptions/{exceptionId}/reopen": {
      "post": {
        "operationId": "reopenPaymentException",
        "parameters": [
          { "$ref": "#/components/parameters/ExceptionId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionExceptionInput"
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PaymentException" },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "409": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/evidence-packs": {
      "post": {
        "operationId": "createEvidencePack",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidencePackInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed evidence generated",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EvidencePack" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ApiError" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" },
          "503": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/v1/evidence-packs/{evidencePackId}/verification": {
      "get": {
        "operationId": "getEvidencePackVerification",
        "parameters": [{ "$ref": "#/components/parameters/EvidencePackId" }],
        "responses": {
          "200": {
            "description": "Historical verification material for this evidence pack",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceVerification"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "404": { "$ref": "#/components/responses/ApiError" },
          "429": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "payops.session_token"
      },
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "headers": {
      "RequestId": {
        "description": "Canonical request UUID",
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
...
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": { "type": "string", "minLength": 16, "maxLength": 1024 }
      },
...
      "ExceptionId": {
        "name": "exceptionId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
...
    },
    "responses": {
...
      "ApiError": {
        "description": "Stable API error",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
...
    },
    "schemas": {
...
      "ExceptionReviewState": {
        "type": "string",
        "enum": [
          "open",
          "assigned",
          "investigating",
          "escalated",
          "resolved",
          "ignored"
        ]
      },
...
    }
  }
}